Refactoring and code simplification.
