+++ tensorflow/compiler/xla/service/gpu/ir_emitter_unnested.cc
@@ -1352,6 +1352,61 @@ Status VerifyBatchNormForThunkEmission(
 
   return Status::OK();
 }
+
+// Determine if we enable the row optimized codegen.  When we have a
+// fusion with only point-wise operations, scalar broadcasting and row
+// broadcasting, we can trigger a kernel that vectorize the row loads.
+// This speed up the kernel, in particular on A100.
+bool EnableRowOptimization(mlir::lmhlo::FusionOp fusion) {
+  bool row_optimized = fusion.getFusionResults().size() == 1 && // Not tested with MOF.
+      absl::c_all_of(GetHloOperands(fusion), [](const mlir::Value& value) {
+          // Only tested when the inputs are row-major. So only enable that case.
+          // Maybe it would works if only the inner dimensions is contiguous.
+          if (auto op = value.getDefiningOp()) {
+            return IsRowMajor(value.getDefiningOp());
+          }
+          // Reuse TypeToShape to not duplicate the layout convertion code.
+          return LayoutUtil::IsMonotonicWithDim0Major(TypeToShape(value.getType()).layout());
+        }) &&
+      // Only tested when the output is row-major.
+      absl::c_all_of(GetOutputOps(fusion), IsRowMajor);
+
+  bool some_row_broadcasting = false;
+  // Check that the operation in the fusion are supported
+  for (mlir::Operation& op : fusion.region().front()) {
+    if (mlir::isa<mlir::memref::TensorLoadOp, mlir::memref::TensorStoreOp,
+        mlir::lmhlo::TerminatorOp, mlir::mhlo::ReturnOp,
+        mlir::mhlo::ConstOp, mlir::lmhlo::ConstOp>(op) ) {
+      continue;
+    }
+    HloOpcode opcode = *MhloToHloOpcode(&op);
+    if(HloInstruction::IsOpElementwise(opcode)) {
+      continue;
+    }
+
+    if (auto broadcast = mlir::dyn_cast<mlir::mhlo::BroadcastInDimOp>(op)) {
+      if (broadcast.broadcast_dimensions().size() == 0) {
+        continue;
+      }
+      std::vector<int64> broadcast_dimensions;
+      for (const llvm::APInt& int_value : broadcast.broadcast_dimensions()) {
+        broadcast_dimensions.push_back(int_value.getSExtValue());
+      }
+
+      auto rank = TypeToShape(broadcast.getResult().getType()).rank();
+      if (broadcast_dimensions.size() == 1 &&
+          broadcast_dimensions.back() == (rank - 1)) {
+        some_row_broadcasting = true;
+        continue;
+      }
+    }
+    row_optimized = false;
+    VLOG(2) << "Row vectorization not enabled due to this op: " << MlirToString(&op);
+    break;
+  }
+  // Trigger only when there is a row broadcasting.
+  return row_optimized &= some_row_broadcasting;
+}
 }  // namespace
 
 Status IrEmitterUnnested::EmitBatchNormThunkFromMlir(MlirEmitterInput input) {
@@ -1800,20 +1855,6 @@ StatusOr<MlirEmitterInput> IrEmitterUnnested::GetMlirEmitterInput(
   return input;
 }
 
-bool IsRowMajor(mlir::Operation* op) {
-  if (auto attr = mlir::GetLayoutFromMlirHlo(op)) {
-    std::vector<int64> minor_to_major;
-    absl::c_transform(
-        attr, std::back_inserter(minor_to_major),
-        std::function<int64(const llvm::APInt&)>(&llvm::APInt::getZExtValue));
-    bool ret = std::is_sorted(minor_to_major.begin(),
-                              minor_to_major.end(), std::greater<int64>());
-    return ret;
-  }
-  // It is row major by default.
-  return true;
-}
-
 // TODO(timshen): update the comment once the HandleFusion code path deleted.
 //
 // This is migrated from IrEmitter::HandleFusion() with IrEmitterUnnested as the
@@ -1889,57 +1930,7 @@ Status IrEmitterUnnested::EmitLoopFusionFromMlir(
     return true;
   }();
 
-  bool row_optimized = fusion.getFusionResults().size() == 1 && // Not tested with MOF.
-      absl::c_all_of(GetHloOperands(fusion), [](const mlir::Value& value) {
-          // Only tested when the inputs are row-major. So only enable that case.
-          // Maybe it would works if only the inner dimensions is contiguous.
-          if (auto op = value.getDefiningOp()) {
-            return IsRowMajor(value.getDefiningOp());
-          }
-          // Reuse TypeToShape to not duplicate the layout convertion code.
-          return LayoutUtil::IsMonotonicWithDim0Major(TypeToShape(value.getType()).layout());
-        }) &&
-      // Only tested when the output is row-major.
-      absl::c_all_of(GetOutputOps(fusion), IsRowMajor);
-
-  bool some_row_broadcasting = false;
-  for (mlir::Operation& op : fusion.region().front()) {
-    if (mlir::isa<mlir::memref::TensorLoadOp, mlir::memref::TensorStoreOp,
-        mlir::lmhlo::TerminatorOp, mlir::mhlo::ReturnOp,
-        mlir::mhlo::ConstOp, mlir::lmhlo::ConstOp>(op) ) {
-      continue;
-    }
-    HloOpcode opcode = *MhloToHloOpcode(&op);
-    if(HloInstruction::IsOpElementwise(opcode)) {
-      continue;
-    }
-
-    if (auto broadcast = mlir::dyn_cast<mlir::mhlo::BroadcastInDimOp>(op)) {
-      std::vector<int64> broadcast_dimensions;
-      if (broadcast.broadcast_dimensions().size() > 0) {
-        for (const llvm::APInt& int_value : broadcast.broadcast_dimensions()) {
-          broadcast_dimensions.push_back(int_value.getSExtValue());
-        }
-      }
-      auto rank = TypeToShape(fusion.getFusionResults()[0].getType()).rank();
-      // The row optimized codepath currently only support scalar and
-      // row broadcasting.
-      if (broadcast_dimensions.size() == 0) {
-        continue;
-      }
-      if (broadcast_dimensions.size() == 1 &&
-          broadcast_dimensions.back() == (rank - 1)) {
-        some_row_broadcasting = true;
-        continue;
-      }
-    }
-    row_optimized = false;
-    VLOG(2) << "Row vectorization not enabled due to this op: " << MlirToString(&op);
-    break;
-  }
-  // Trigger only when there is a row broadcasting.
-  row_optimized &= some_row_broadcasting;
-
+  bool row_optimized = EnableRowOptimization(fusion);
   Shape element_shape = context.output_shapes[0];
   TF_ASSIGN_OR_RETURN(LaunchDimensions launch_dimensions,
                       CalculateLaunchDimensions(
